implemented alinea h ish implemented a lot of mapper functions several fixes

+++ F2/src/main/java/isel/sisinf/grp02/data_mappers/ICondutorMapper.java
@@ -5,5 +5,5 @@ import isel.sisinf.grp02.repositories.IRepository;
 
 import java.util.Collection;
 
-public interface ICondutorMapper extends IDataMapper<Condutor, Long> {
+public interface ICondutorMapper extends IDataMapper<Condutor, Integer> {
 }
